Move graphicalRepresentation
