+++ assets/src/pages/graphsPage.js
@@ -1,12 +1,9 @@
 import { initCore } from 'src/lib/init.js'
 import renderComponents from 'src/lib/renderComponents.js'
-import 'vis/dist/vis.css'
-import graphicalRepresentation from '@/graphicalRepresentation.vue'
 import graphsPage from '@/graphs/graphsPage.vue'
 
 
 initCore()
 renderComponents({
-    'graphical-representation': graphicalRepresentation,
     'graphs-page': graphsPage
 })
